requirements.txt
<<<<<<< HEAD
python-telegram-bot==20.7
requests==2.31.0
firebase-admin==6.4.0
python-dotenv==1.0.1
openai==1.12.0
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scikit-learn>=1.4,<2.0
feedparser==6.0.11
telethon ==1.42.0
=======
annotated-types==0.7.0
anyio==4.12.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
feedparser==6.0.11
firebase-admin==6.4.0
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.25.2
idna==3.11
joblib==1.5.3
msgpack==1.1.2
numpy==1.26.4
openai==1.12.0
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==2.3.3
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2025.2
requests==2.31.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
Telethon==1.42.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
>>>>>>> eaacaf7 (deploy)
